Sin and Codex: adopt fractional Viewer camera and Party motion
Carry Double through the existing Viewer and Sin Star I camera, actor and renderer owners after the published P0, D1 and D2 milestones. This publishes implemented production adoption and validated automated delivery; the complete native/Chrome manual acceptance remains explicitly open in the single checkpoint.

Implementation:
- Add the focused shared PrecisionCamera3D controller/framing/projection owner and use it in ViewerCamera and the Sin Star I preview without changing scene scale or the 18-degree/second preview rate.
- Evaluate actual Party approach/return from unchanged integer action time, place the same Character3D actors, and retain fractions in current-pose sockets, equipment pivots, glow, ribbons, fire, lightning, GPU spawn staging and light leases.
- Preserve actor pools, rendering ownership, event/noise order, IDs, budgets, authored units and integral saved calibration. Keep Program.smile's accepted coordinator; add only typed wiring. Add a bounded shared cubic visual source/project.
- Correct integration defects in Web pivot initialization/payload reads, unit camera-up, native Double/music CRT linkage, fieldless debug views and compile-time Enum constants versus runtime Web Number bounds.

Validation and delivery:
- Native/generated-Web camera, Character3D normal/fallback, Lightning, VFX staging/commit/light and Viewer hardening fixtures pass. The real Party owner reaches captured render submissions at fractional approach/return positions and passes equal elapsed-time partitions on both targets.
- Full normal smoke checks pass through focused continuation after repaired failures; Doctor is explicitly skipped. 319 shared tests, 13 formatter tests, 430 tracked-source style checks and the focused new-source check pass. Final build has no NU1503; current assets and artifact metadata verify.
- Build Viewer native/Full Web, Sin Star I native/Web and the cubic visual native/Web. Install VSIX 2.0.61 without closing Visual Studio, and verify 35 installed payload hashes against the built archive. Export both live character calibrations: Arin 24 keys and Orin zero keys, unchanged canonical content.

Limits and remaining acceptance:
- Native Double is binary64; GPU acceptance remains float32. Package format 7 requires source rebuilding of older packages; integer Number APIs retain explicit boundaries.
- Chrome visibly renders all three selected consumers and the preview responds to pan/zoom/floor controls. The Viewer shows roughly one-second frame scheduling and some brief input is missed; no smoothness or complete audible/manual pass is claimed.
- Native app APIs are disabled in this session despite the user's desktop permission. Preserve the pre-existing Viewer and finish native/foreground camera, Party, pause/scrub, attachment/floor/audio, curve and installed VS breakpoint/hover checks when control is available. No unrelated suspended feature is started.

+++ src/Smile.NativeRuntime/graphics/graphics3d_directx.cpp
@@ -1901,8 +1901,8 @@ static long long smile_3d_create_gpu_particle_system(unsigned int capacity,
 }
 
 static int smile_3d_stage_gpu_particle_kinematics(SmileGpuParticleSystem3D* system,
-    long long slot, long long x, long long y, long long z,
-    long long velocity_x, long long velocity_y, long long velocity_z)
+    long long slot, double x, double y, double z,
+    double velocity_x, double velocity_y, double velocity_z)
 {
     SmileGpuParticleState3D* state;
     if (slot < 0 || slot >= system->capacity ||
@@ -2825,8 +2825,8 @@ static int smile_3d_set_directional(long long x, long long y, long long z,
     return 1;
 }
 
-static int smile_3d_set_local_light(long long slot, long long type, long long x, long long y,
-    long long z, long long red, long long green, long long blue, long long intensity, long long range)
+static int smile_3d_set_local_light(long long slot, long long type, double x, double y,
+    double z, long long red, long long green, long long blue, long long intensity, long long range)
 {
     SmileLocalLight3D* light;
     if (slot < 0 || slot >= SMILE_3D_MAX_LOCAL_LIGHTS || type < 0 || type > 2)
@@ -9487,7 +9487,8 @@ static long long smile_3d_gpu_particle_system_command(long long operation,
         return 0;
     }
     if (operation == 2)
-        return smile_3d_stage_gpu_particle_kinematics(system, c, d, e, f, g, h, i);
+        return smile_3d_stage_gpu_particle_kinematics(system, c, (double)d, (double)e,
+            (double)f, (double)g, (double)h, (double)i);
     if (operation == 3)
         return smile_3d_stage_gpu_particle_visual(system, c, d, e, f, g, h, i, j);
     if (operation == 4) return smile_3d_commit_gpu_particle_spawn(system, c, d, e);
@@ -10028,7 +10029,7 @@ extern "C" long long smile_renderer3d_command(long long command,
         case SMILE_3D_SET_DIRECTIONAL_LIGHT:
             return smile_3d_set_directional(a, b, c, d, e, f, g);
         case SMILE_3D_SET_LOCAL_LIGHT:
-            return smile_3d_set_local_light(a, b, c, d, e, f, g, h, i, j);
+            return smile_3d_set_local_light(a, b, (double)c, (double)d, (double)e, f, g, h, i, j);
         case SMILE_3D_SET_SPOT_CONE: return smile_3d_set_spot_cone(a, b, c, d, e, f);
         case SMILE_3D_PBR_TEXTURE_VALUE: return smile_3d_pbr_texture_value(texture = smile_3d_texture(a), b);
         case SMILE_3D_PBR_MATERIAL_VALUE: return smile_3d_pbr_material_value(material = smile_3d_material(a), b);
